Refactor sygvdx-hegvdx tests
This commit updates the error bounds and fixes bugs in the sygvdx-hegvdx
tests.  It also adds code to allow selecting whether the tests should
enforce equality on the number of reference and computed eivenvalues
(default is yes); and whether the tests should use new or legacy
error bounds (default is to use new error bounds).